+++ lib/favorite/provider/favorite_provider.dart
@@ -1,10 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:knu_helper/common/model/cursor_pagination_model.dart';
-import 'package:knu_helper/common/utils/data_utils.dart';
 import 'package:knu_helper/favorite/repository/favorite_repository.dart';
-import 'package:knu_helper/notice/database/drift_database.dart';
+
 import 'package:knu_helper/notice/model/notice_model.dart';
-import 'package:knu_helper/notice/repository/notice_repository.dart';
 
 final favoriteProvider =
     StateNotifierProvider<FavoriteStateNotifier, List<NoticeModel>>((ref) {
