+++ lib/notice/provider/notice_provider.dart
@@ -1,12 +1,12 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:knu_helper/common/model/cursor_pagination_model.dart';
 import 'package:knu_helper/favorite/repository/favorite_repository.dart';
-import 'package:knu_helper/notice/database/drift_database.dart';
 import 'package:knu_helper/notice/model/notice_model.dart';
 import 'package:knu_helper/notice/model/site_color.dart';
 import 'package:knu_helper/notice/repository/notice_repository.dart';
 
 import '../../all/provider/user_site_provider.dart';
+import '../../common/database/drift_database.dart';
 
 final noticeProvider =
     StateNotifierProvider<NoticeStateNotifier, CursorPaginationBase>((ref) {
